rename scripts file, test report, test stake, scaffolding
